+++ lib/dlms-cosem/include/dlms/cosem/simple_objects.hpp
@@ -1460,6 +1460,54 @@ private:
   CosemAccessRights rights_;
 };
 
+class CosemMBusSlavePortSetupObject : public ICosemObject
+{
+public:
+  static const std::uint8_t MaxSupportedVersion = 0u;
+
+  CosemMBusSlavePortSetupObject(
+    const CosemLogicalName& logicalName,
+    const CosemByteBuffer& defaultBaud,
+    const CosemByteBuffer& availableBaud,
+    const CosemByteBuffer& status,
+    const CosemByteBuffer& mbusPortReference,
+    AttributeAccessMode mutableAccess);
+  CosemMBusSlavePortSetupObject(
+    const CosemLogicalName& logicalName,
+    const CosemByteBuffer& defaultBaud,
+    const CosemByteBuffer& availableBaud,
+    const CosemByteBuffer& status,
+    const CosemByteBuffer& mbusPortReference,
+    AttributeAccessMode mutableAccess,
+    std::uint8_t version);
+
+  CosemObjectDescriptor Descriptor() const;
+  CosemAccessRights AccessRights() const;
+  CosemStatus ReadAttribute(
+    std::uint8_t attributeId,
+    CosemByteBuffer& output) const;
+  CosemStatus WriteAttribute(
+    std::uint8_t attributeId,
+    const CosemByteBuffer& input);
+  CosemStatus InvokeMethod(
+    std::uint8_t methodId,
+    const CosemByteBuffer& input,
+    CosemByteBuffer& output);
+
+  const CosemByteBuffer& DefaultBaud() const;
+  const CosemByteBuffer& AvailableBaud() const;
+  const CosemByteBuffer& Status() const;
+  const CosemByteBuffer& MBusPortReference() const;
+
+private:
+  CosemObjectDescriptor descriptor_;
+  CosemByteBuffer defaultBaud_;
+  CosemByteBuffer availableBaud_;
+  CosemByteBuffer status_;
+  CosemByteBuffer mbusPortReference_;
+  CosemAccessRights rights_;
+};
+
 enum class CosemClockBase
 {
   NotDefined = 0,